[refactor]: rename point module class file names snake_case
