+++ spatialetl/point/io/ascii/default_time_point_reader.py
@@ -28,7 +28,7 @@ import numpy as np
 
 import pandas
 
-from spatialetl.point.TimeMultiPoint import TimeMultiPoint
+from spatialetl.point.time_multi_point import TimeMultiPoint
 from spatialetl.point.io.multi_point_reader import MultiPointReader
 from spatialetl.utils.logger import logging
