+++ spatialetl/point/io/ascii/default_time_point_writer.py
@@ -28,7 +28,7 @@ from datetime import datetime
 import numpy as np
 import pandas
 
-from spatialetl.point.TimeMultiPoint import TimeMultiPoint
+from spatialetl.point.time_multi_point import TimeMultiPoint
 from spatialetl.point.io.multi_point_writer import MultiPointWriter
 from spatialetl.utils.VariableDefinition import VariableDefinition
 from spatialetl.utils.logger import logging
