+++ spatialetl/point/level_multi_point.py
@@ -25,7 +25,7 @@ from __future__ import division, print_function, absolute_import
 import numpy as np
 
 from spatialetl.operator.interpolator.interpolator_core import vertical_interpolation
-from spatialetl.point.MultiPoint import MultiPoint
+from spatialetl.point.multi_point import MultiPoint
 from spatialetl.utils.logger import logging
