+++ spatialetl/point/tests/test_time_multi_point.py
@@ -20,7 +20,7 @@
 # SOFTWARE.
 from unittest import TestCase
 from datetime import datetime
-from spatialetl.point.TimeMultiPoint import TimeMultiPoint
+from spatialetl.point.time_multi_point import TimeMultiPoint
 from spatialetl.point.io.ascii.default_time_point_reader import DefaultTimePointReader
