+++ spatialetl/providers/common/netcdf/spatialetl/point/io/netcdf/DefaultTimeLevelMultiPointReader.py
@@ -25,7 +25,7 @@ from __future__ import division, print_function, absolute_import
 import numpy as np
 from netCDF4 import Dataset, num2date
 
-from spatialetl.point.TimeMultiPoint import TimeMultiPoint
+from spatialetl.point.time_multi_point import TimeMultiPoint
 from spatialetl.point.io.multi_point_reader import MultiPointReader
 from spatialetl.utils.VariableDefinition import VariableDefinition
