+++ spatialetl/providers/common/netcdf/spatialetl/point/io/netcdf/DefaultWriter.py
@@ -27,9 +27,9 @@ from netCDF4 import Dataset
 from netCDF4 import date2num
 from numpy import float32, float64, int32
 
-from spatialetl.point.LevelMultiPoint import LevelMultiPoint
-from spatialetl.point.TimeLevelMultiPoint import TimeLevelMultiPoint
-from spatialetl.point.TimeMultiPoint import TimeMultiPoint
+from spatialetl.point.level_multi_point import LevelMultiPoint
+from spatialetl.point.time_level_multi_point import TimeLevelMultiPoint
+from spatialetl.point.time_multi_point import TimeMultiPoint
 from spatialetl.point.io.multi_point_writer import MultiPointWriter
 from spatialetl.utils.VariableDefinition import VariableDefinition
 from spatialetl.utils.logger import logging
